requirements.txt
asgiref==3.4.1
Brotli==1.0.9
Django==3.2.5
django-ckeditor==6.1.0
django-compression-middleware==0.4.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.12.0
django-storages==1.11.1
Pillow==8.3.1
psycopg2-binary==2.9.1
python-memcached==1.59
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
zstandard==0.15.2
